Implement order creation functionality
Enhanced order creation process in the 'order-service' by adding product and coupon locking functionality, as well as validation checks. In the 'product-service', added the ability for confirming order cart items which get removed from the cart once confirmed. Also fixed an endpoint in 'coupon-service' by correcting the mapping for locking coupon records. This effort also involved the creation of several new classes and modifications to existing ones across services to support the implementation.